requirements_v2.txt
hydra-core==1.3.2
ipython==8.4.0
librosa==0.8.1
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.3.0
pytorch_lightning==1.9.0
scipy==1.7.3
torch==1.14.0
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
pytorch-lightning>=1.6.0,<2.0.0